+++ src/countries/al.ts
@@ -12,6 +12,7 @@ export const al: Country = {
 		pt: 'Albânia',
 		tr: 'Arnavutluk',
 		fa: 'آلبانی',
+		uk: 'Албанія',
 		xa: 'Himni i Flamurit',
 		xt: 'Himni i Flamurit',
 	},