+++ src/countries/de.ts
@@ -12,6 +12,7 @@ export const de: Country = {
 		pt: 'Alemanha',
 		tr: 'Almanya',
 		fa: 'آلمان',
+		uk: 'Німеччина',
 		xa: 'Das Lied der Deutschen',
 		xt: 'Das Lied der Deutschen',
 	},