+++ src/countries/ua.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const ua: Country = {
+	code: 'ua',
+	name: {
+		en: 'Ukraine',
+		ar: 'أوكرانيا',
+		de: 'Ukraine',
+		sv: 'Ukraina',
+		da: 'Ukraine',
+		sq: 'Ukraina',
+		pt: 'Ucrânia',
+		tr: 'Ukrayna',
+		fa: 'اوکراین',
+		uk: 'Україна',
+		xa: 'Ще не вмерла України і слава, і воля',
+		xt: 'Ще не вмерла України і слава, і воля',
+	},
+	flag: '🇺🇦',
+}