+++ src/countries/us.ts
@@ -12,6 +12,7 @@ export const us: Country = {
 		pt: 'Estados Unidos da América',
 		tr: 'Amerika Birleşik Devletleri',
 		fa: 'ایالات متحده آمریکا',
+		uk: 'Сполучені Штати Америки',
 		xa: 'The Star-Spangled Banner',
 		xt: 'The Star-Spangled Banner',
 	},